Implement the required changes
